% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_sum_cpp}
\alias{calc_sum_cpp}
\title{Calculate the sum}
\usage{
calc_sum_cpp(values)
}
\arguments{
\item{values}{the values to sum}
}
\value{
the sum
}
\description{
Calculate the sum
}
\author{
Richèl J.C. Bilderbeek
}

