% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_proper_divisors}
\alias{get_proper_divisors}
\title{Get proper divisors}
\usage{
get_proper_divisors(n)
}
\arguments{
\item{n}{the number to euvfwekyb}
}
\value{
the proper divisors
}
\description{
Get proper divisors
}
\author{
Richèl J.C. Bilderbeek
}

